Download Files from an AWS S3 Bucket to a Local Drive

Copies files from your AWS S3 bucket to a local drive

rwa_download_files(
  file_list,
  directory,
  bucket_name,
  silent = FALSE,
  ask = TRUE,
  aws_access_key_id = Sys.getenv("AWS_ACCESS_KEY_ID"),
  aws_secret_access_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  region = Sys.getenv("AWS_REGION", "ca-central-1")
)

Arguments

file_list

A character vector of file(s).

directory

A string of the local directory to copy the files to. The directory must already exist.

bucket_name

A string of the AWS S3 bucket name.

silent

A flag (by default FALSE) to toggle the progress bar on and off. Set to TRUE to turn off progress bar.

ask

A flag (by default TRUE) to toggle on and off being asked about potentially overwriting files present in the directory argument. Set to FALSE to not be asked.

aws_access_key_id

A string of your AWS user access key ID. The default is the environment variable named AWS_ACCESS_KEY_ID.

aws_secret_access_key

A string of your AWS user secret access key. The default is the environment variable named AWS_SECRET_ACCESS_KEY.

region

A string of the AWS region. The default is the environment variable named AWS_REGION.

Details

The user is prompted about overwriting files if the directory is not completely empty. The function doesn't check which files are present only if its empty or not. Use at your own risk as it will overwrite any files with matching file paths to the files in the S3 bucket.

Examples

if (FALSE) { # \dontrun{
# Download a single file
rwa_download_files(
  file_list = "shiny-upload/image/2021-06-30_16-06-10_test_04d855/uploaded_file.jpeg",
  directory = "purple-lake",
  bucket_name = "purple-lake-poissonconsulting"
)

# Download multiple files
files <- c(
  "shiny-upload/image/2021-06-30_16-06-10_test_04d855/uploaded_file.jpeg",
  "shiny-upload/image/2021-06-30_16-06-10_test_04d855/input_data.csv"
)
rwa_download_files(
  file_list = files,
  directory = "purple-lake",
  bucket_name = "purple-lake-poissonconsulting"
)

# Turn progress bar off when downloading
rwa_download_files(
  file_list = "shiny-upload/image/2021-06-30_16-06-10_test_04d855/uploaded_file.jpeg",
  directory = "purple-lake",
  bucket_name = "purple-lake-poissonconsulting",
  silent = TRUE
)

# Do not ask about potentially overwriting files in the given directory
rwa_download_files(
  file_list = "shiny-upload/image/2021-06-30_16-06-10_test_04d855/uploaded_file.jpeg",
  directory = "purple-lake",
  bucket_name = "purple-lake-poissonconsulting",
  ask = FALSE
)

# Enter AWS credentials directly into the function
rwa_download_files(
  file_list = "shiny-upload/image/2021-06-30_16-06-10_test_04d855/uploaded_file.jpeg",
  directory = "purple-lake",
  bucket_name = "purple-lake-poissonconsulting",
  aws_access_key_id = "AHSGYWKJDIUAHDSJ",
  aws_secret_access_key = "8HYGD54//hgdx^785809",
  region = "us-east-1"
)
} # }