List Files in an AWS S3 Bucket

Obtain a list of files within an AWS S3 Bucket

rwa_list_files(
  bucket_name,
  max_request_size = 1000,
  pattern = ".*",
  silent = FALSE,
  aws_access_key_id = Sys.getenv("AWS_ACCESS_KEY_ID"),
  aws_secret_access_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  region = Sys.getenv("AWS_REGION", "ca-central-1")
)

Arguments

bucket_name

A string of the AWS S3 bucket name.

max_request_size

A whole number (by default 1000) indicating the maximum number of files to be returned.

pattern

A string (by default ".*") containing a regular expression. The default value will return all files retrieved from AWS.

silent

A flag (by default FALSE) to silence messages about number of files returned. Set to TRUE to silence messages.

aws_access_key_id

A string of your AWS user access key ID. The default is the environment variable named AWS_ACCESS_KEY_ID.

aws_secret_access_key

A string of your AWS user secret access key. The default is the environment variable named AWS_SECRET_ACCESS_KEY.

region

A string of the AWS region. The default is the environment variable named AWS_REGION.

Value

Character vector of files names.

Details

The AWS API only allows a max of 1000 objects to return per request. The max_request_size argument is for creating multiple requests to the API so you are able to get a list of all the files if there are more then 1000 objects in your bucket. If max_request_size is over 1000 it will make multiple request to AWS until max_request_size is reached or all the files are listed.

The pattern argument is a regex that is applied after all the files are obtained from AWS. The default value will ensure all files retrieved from AWS are listed. If other values are given you may get less then the value for max_request_size as it will only return the objects that match the pattern.

Review the messages if you are unsure if you should increase max_request size.

Examples

if (FALSE) { # \dontrun{
# List all files in the bucket (up to a max of 1000)
rwa_list_files(
  bucket_name = "purple-lake-poissonconsulting"
)

# Add regex to only list pdfs from June 30, 2020
rwa_list_files(
  bucket_name = "purple-lake-poissonconsulting",
  pattern = "pdf.*2021-06-30"
)

# List all files in the bucket up to a max of 2000
rwa_list_files(
  bucket_name = "purple-lake-poissonconsulting",
  max_request_size = 2000
)

# Turn off help messages regarding total number of files returned
rwa_list_files(
  bucket_name = "purple-lake-poissonconsulting",
  silent = TRUE
)

# Enter AWS credentials directly into the function
rwa_list_files(
  bucket_name = "purple-lake-poissonconsulting",
  aws_access_key_id = "AHSGYWKJDIUAHDSJ",
  aws_secret_access_key = "8HYGD54//hgdx^785809",
  region = "us-east-1"
)
} # }